style: publicResource package removal, -Moved TestApi, Login, Refresh, User Controllers to parent api package, -Removed publicResource child package of api package.
