+++ src/main/java/pl/simpleascoding/tutoringplatform/api/RefreshController.java
@@ -1,28 +1,28 @@
-package pl.simpleascoding.tutoringplatform.api.publicResource;
+package pl.simpleascoding.tutoringplatform.api;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import pl.simpleascoding.tutoringplatform.dto.CredentialsDTO;
 import pl.simpleascoding.tutoringplatform.service.security.SecurityService;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.Map;
 
+import static pl.simpleascoding.tutoringplatform.common.SecurityFinals.AUTHORIZATION;
+
 @RestController
-@RequestMapping("api/v1/login")
+@RequestMapping("api/v1/refresh")
 @RequiredArgsConstructor
-class LoginController {
+class RefreshController {
 
     private final SecurityService service;
 
     @PostMapping
-    public ResponseEntity<Map<String, String>> login(@RequestBody CredentialsDTO dto, HttpServletRequest request) {
-        return new ResponseEntity<>(service.login(dto, request.getRequestURL().toString()), HttpStatus.OK);
+    ResponseEntity<Map<String, String>> refresh(HttpServletRequest request) {
+        return new ResponseEntity<>(service.refresh(request.getHeader(AUTHORIZATION), request.getRequestURL().toString()), HttpStatus.OK);
     }
 
 }
